Move ScooterState to domain:
