+++ lib/domain/scooter_state.dart
