+++ lib/scooter_service.dart
@@ -11,7 +11,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:unustasis/domain/scooter_power_state.dart';
 import 'package:unustasis/flutter/blue_plus_mockable.dart';
 import 'package:unustasis/home_screen.dart';
-import 'package:unustasis/scooter_state.dart';
+import 'package:unustasis/domain/scooter_state.dart';
 
 const bootingTimeSeconds = 25;
 const keylessCooldownSeconds = 60;
