+++ lib/stats/range_section.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:unustasis/scooter_service.dart';
-import 'package:unustasis/scooter_state.dart';
+import 'package:unustasis/domain/scooter_state.dart';
 import 'package:unustasis/scooter_visual.dart';
 import 'package:unustasis/stats/battery_section.dart';
